add file UserModel test

+++ models/UserModel.php
@@ -1,11 +1,15 @@
 <?php
 
+use function PHPUnit\Framework\isEmpty;
+
 require_once 'BaseModel.php';
 
-class UserModel extends BaseModel {
-  public static function getInstance() {
-    if (self::$_instance !== null){
-        return self::$_instance;
+class UserModel extends BaseModel
+{
+  public static function getInstance()
+  {
+    if (self::$_instance !== null) {
+      return self::$_instance;
     }
     self::$_instance = new self();
     return self::$_instance;
@@ -25,20 +29,61 @@ class UserModel extends BaseModel {
     return $user;
   }
   //--------------------------------------------------------------
+  function checkUserExist($id)
+  {
+    if (is_null($id) || is_string($id) || !is_numeric($id) || empty($id)) {
+      return "error";
+    }
+    $sql = 'SELECT * FROM users WHERE  `id` =' . $id;
+    $user = $this->select($sql);
+    if (empty($user)) {
+      return false;
+    }
+    if ($user[0]["id"] == $id) {
+      return true;
+    }
+  }
+  //todo--------------------------------------------------------------
   public function deleteUserById($id)
   {
-    $sql = 'DELETE FROM users WHERE id = ' . $id;
-    return $this->delete($sql);
+    if (is_null($id) || is_string($id) || !is_numeric($id) || empty($id)) {
+      return "error";
+    }
+    if ($this->checkUserExist($id) == true) {
+      $sql = 'DELETE FROM users WHERE id = ' . $id;
+      $check = $this->delete($sql);
+      return true;
+    } else {
+      return "User doesn't exis";
+    }
   }
   //--------------------------------------------------------------
   public function updateUser($input)
   {
-    $sql = 'UPDATE users SET 
-                  name = "' . $input['name'] . '", 
-                  password="' . md5($input['password']) . '"
-                WHERE id = ' . $input['id'];
-    $user = $this->update($sql);
-    return $user;
+
+    if (
+      is_numeric(($input['id'])) == false || $input['id'] == 'e'
+      || is_string($input['name']) == false || is_string($input['fullname']) == false || is_string($input['email']) == false
+      || is_string($input['type']) == false || is_string($input['password']) == false
+      || strlen($input['name']) == 0 || strlen($input['fullname']) == 0 || strlen($input['email']) == 0
+      || strlen($input['type']) == 0 || strlen($input['password']) == 0
+    ) {
+      return "error";
+    }
+
+    if ($this->checkUserExist(($input['id'])) == true) {
+      $sql = 'UPDATE users SET 
+      name = "' . mysqli_real_escape_string(self::$_connection, $input['name'])  . '", 
+      fullname="' . ($input['fullname']) . '",
+      email="' . ($input['email']) . '",
+      password="' . (md5($input['password'])) . '",
+      type="' . $input['type'] . '"
+      WHERE id = ' . ($input['id']);
+      $user = $this->update($sql);
+      return true;
+    } else {
+      return "User not exist";
+    }
   }
   //--------------------------------------------------------------
   public function insertUser($input)
@@ -59,17 +104,16 @@ class UserModel extends BaseModel {
     //Keyword
     $users = null;
     if (!empty($params['keyword'])) {
-        $stmt = self::$_connection->prepare("SELECT * FROM users WHERE name LIKE CONCAT('%',?,'%')");
-        if($stmt) {
-            $stmt->bind_param("s", $params['keyword']);
-            $stmt->execute();
-            $users = array();
-            $users = $stmt->get_result()->fetch_all(MYSQLI_ASSOC);
-        }
-    
+      $stmt = self::$_connection->prepare("SELECT * FROM users WHERE name LIKE CONCAT('%',?,'%')");
+      if ($stmt) {
+        $stmt->bind_param("s", $params['keyword']);
+        $stmt->execute();
+        $users = array();
+        $users = $stmt->get_result()->fetch_all(MYSQLI_ASSOC);
+      }
     } else {
-        $sql = 'SELECT * FROM users';
-        $users = $this->select($sql);
+      $sql = 'SELECT * FROM users';
+      $users = $this->select($sql);
     }
 
     return $users;
